High scores are now saved for each level in NSUserDefaults.

+++ MonkeySwing/MyScene.h
@@ -12,6 +12,6 @@
 @interface MyScene : SKScene <SKPhysicsContactDelegate>
 
 @property (nonatomic, strong) PhysicsParameters *physicsParameters;
-@property (nonatomic) NSInteger levelNumber;
+@property (nonatomic) int levelNumber;
 
 @end
